requirements.txt
websockets
requests
python-dotenv

